+++ database/migrations/20250208141539_create_table_wallets.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE wallets (
     id INTEGER AUTO_INCREMENT,
     user_id INTEGER NOT NULL,
-    balance INTEGER NOT NULL,
+    balance DECIMAL(15, 2) NOT NULL,
     status TINYINT(1) NOT NULL COMMENT '1 : active | 2 : inactive | 3 : suspend',
     created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
     updated_at DATETIME DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
